Split out BaseConfig from OCPNConfig
OCPNConfig depends on the GUI so it cannot be used in non-GUI
model code. Split it into BaseConfig without any GUI dependencies
and keep the remaining GUI-dependent stuff in OCPNPlatform.
