+++ src/PluginHandler.cpp
@@ -34,9 +34,11 @@
 #include <unordered_map>
 #include <set>
 
-#include <wx/jsonreader.h>
-#include <wx/string.h>
+//#include <wx/jsonreader.h>
+#include <wx/dir.h>
 #include <wx/file.h>
+#include <wx/string.h>
+#include <wx/window.h>
 #include <wx/uri.h>
 
 #include <archive.h>
@@ -55,7 +57,7 @@ typedef __LA_INT64_T la_int64_t;  //  "older" libarchive versions support
 #include "logger.h"
 #include "navutil.h"
 #include "gui_lib.h"
-#include "OCPNPlatform.h"
+#include "BasePlatform.h"
 #include "ocpn_utils.h"
 #include "PluginHandler.h"
 #include "plugin_cache.h"
@@ -72,13 +74,12 @@ static std::string SEP("/");
 #define F_OK 0
 #endif
 
-extern OCPNPlatform* g_Platform;
+extern BasePlatform* g_Platform;
 extern PlugInManager* g_pi_manager;
 extern wxString g_winPluginDir;
 extern MyConfig* pConfig;
-extern OCPNPlatform* g_Platform;
 extern bool g_bportable;
-extern MyFrame* gFrame;
+extern wxWindow* gFrame;
 
 extern wxString g_compatOS;
 extern wxString g_compatOsVersion;
@@ -323,6 +324,8 @@ std::string PluginHandler::fileListPath(std::string name) {
   std::transform(name.begin(), name.end(), name.begin(), ::tolower);
   return pluginsConfigDir() + SEP + name + ".files";
 }
+
+PluginHandler::PluginHandler() {}
 
 bool PluginHandler::isCompatible(const PluginMetadata& metadata, const char* os,
                                  const char* os_version) {
@@ -1146,19 +1149,23 @@ bool PluginHandler::installPluginFromCache(PluginMetadata plugin) {
     bool bOK = installPlugin(plugin, cacheFile);
     if (!bOK) {
       wxLogWarning("Cannot install tarball file %s", cacheFile.c_str());
+      evt_download_failed.notify(cacheFile);
+      /** FIXME
       wxString message = _("Please check system log for more info.");
       OCPNMessageBox(gFrame, message, _("Installation error"),
                      wxICON_ERROR | wxOK | wxCENTRE);
-
+      **/ 
       return false;
     }
 
     wxString message;
     message.Printf("%s %s\n", plugin.name.c_str(), plugin.version.c_str());
+    evt_download_ok.notify(message.ToStdString());
+/** FIXME
     message += _(" successfully installed from cache");
     OCPNMessageBox(gFrame, message, _("Installation complete"),
                    wxICON_INFORMATION | wxOK | wxCENTRE);
-
+**/
     return true;
   }
