+++ src/catalog_parser.cpp
@@ -23,6 +23,9 @@
  */
 
 #include <cstring>
+#include <fstream>
+#include <iterator>
+#include <sstream>
 
 #include <wx/log.h>
 
@@ -33,12 +36,8 @@
 #include "catalog_parser.h"
 #include "ocpn_utils.h"
 #include "catalog_handler.h"
-#include "Downloader.h"
 #include "pugixml.hpp"
 
-#include <fstream>
-#include <sstream>
-#include <iterator>
 
 bool ParseCatalog(const std::string xml, catalog_ctx* ctx) {
   bool ok = true;
