+++ src/plugin_cache.cpp
@@ -27,15 +27,16 @@
 #include <wx/filename.h>
 #include <wx/filefn.h>
 
+#include "BasePlatform.h"
 #include "plugin_cache.h"
 #include "ocpn_utils.h"
-#include "OCPNPlatform.h"
+
 
 #ifdef __OCPN__ANDROID__
 #include "androidUTIL.h"
 #endif
 
-extern OCPNPlatform* g_Platform;
+extern BasePlatform* g_Platform;
 
 static std::string cache_path() {
   wxFileName path;
