+++ src/pluginmanager.cpp
@@ -602,33 +602,6 @@ static void run_update_dialog(PluginListPanel *parent, PlugInContainer *pic,
   // g_pi_manager->GetListPanelPtr()->SelectByName(name);
 }
 
-wxString GetPluginDataDir(const char *plugin_name) {
-  static const wxString sep = wxFileName::GetPathSeparator();
-
-  wxString datadirs = g_Platform->GetPluginDataPath();
-  wxLogMessage(_T("PlugInManager: Using data dirs from: ") + datadirs);
-  wxStringTokenizer dirs(datadirs, ";");
-  while (dirs.HasMoreTokens()) {
-    wxString dir = dirs.GetNextToken();
-    wxFileName tryDirName(dir);
-    wxDir tryDir;
-    if (!tryDir.Open(tryDirName.GetFullPath())) continue;
-    wxString next;
-    bool more = tryDir.GetFirst(&next);
-    while (more) {
-      if (next == plugin_name) {
-        next = next.Prepend(tryDirName.GetFullPath() + sep);
-        wxLogMessage(_T("PlugInManager: using data dir: %s"), next);
-        return next;
-      }
-      more = tryDir.GetNext(&next);
-    }
-    tryDir.Close();
-  }
-  wxLogMessage(_T("Warnińg: no data directory found, using \"\""));
-  return "";
-}
-
 //    Some static helper funtions
 //    Scope is local to this module
 
@@ -1055,6 +1028,14 @@ PlugInManager::~PlugInManager() {
 #endif
 #endif
 }
+
+
+/**
+ * Set up actions to perform for messages generated by PluginHandlers's
+ * evt_foo.notify() calls.
+ */
+wxDEFINE_EVENT(EVT_DOWNLOAD_FAILED, wxCommandEvent);
+wxDEFINE_EVENT(EVT_DOWNLOAD_OK, wxCommandEvent);
 
 /**
  *
